Added ExampleRunner trait to execute examples #15
